# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/tenants": {
      "post": {
        "operationId": "TenancyController_register",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated ULID key; replays return the original response",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 26,
              "maxLength": 26,
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRegistrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key, or invalid body",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Owner email already registered (duplicate-email)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Idempotency key reused with a different payload (idempotency-key-reuse)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Registers a tenant with its Owner user",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenants/sign-in": {
      "post": {
        "operationId": "TenancyController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unknown email or wrong password (unauthenticated)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Verifies the password and issues a short-lived session token",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenants/{tenantId}/warehouses": {
      "post": {
        "operationId": "TenancyController_createWarehouse",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated ULID key; replays return the original response",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 26,
              "maxLength": 26,
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key, or invalid body",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Warehouse code already exists (duplicate-warehouse-code names the code)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Idempotency key reused with a different payload (idempotency-key-reuse)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Creates a warehouse (code unique per tenant)",
        "tags": [
          "tenancy"
        ]
      },
      "get": {
        "operationId": "TenancyController_listWarehouses",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque keyset cursor from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed cursor",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lists warehouses (keyset cursor pagination)",
        "tags": [
          "tenancy"
        ]
      }
    },
...
    "/tenants/{tenantId}/warehouses/{warehouseId}/zones/{zoneId}/bins/grid": {
      "post": {
        "operationId": "TenancyController_generateBinGrid",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated ULID key; replays return the original response",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 26,
              "maxLength": 26,
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateBinsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BinGridResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key, invalid body, or a descending aisle range (validation-failed)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Warehouse or zone does not exist in this tenant (not-found)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "A generated code collides (duplicate-bin-code names the first conflicting code; nothing committed)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Grid exceeds 500 bins (grid-too-large), or idempotency-key-reuse",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mass-creates bins in one zone from an aisle/bay/level grid (≤ 500, one transaction)",
        "tags": [
          "tenancy"
        ]
      }
    },
...
    "/tenants/{tenantId}/setup-checklist": {
      "get": {
        "operationId": "TenancyController_setupChecklist",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupChecklistResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Computed onboarding checklist (warehouse, bins, catalog, users — catalog/users pending)",
        "tags": [
          "tenancy"
        ]
      }
    },
...
  },
  "info": {
    "title": "WMS API",
    "description": "Warehouse Management System backend — modular monolith. This document is the contract: web and mobile clients generate typed clients from it (AD-8).",
    "version": "1.0.0",
    "contact": {
      "name": "WMS",
      "url": "https://wms.example.com",
      "email": "api@wms.example.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wms.example.com/license"
    }
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Priya Spices Pvt Ltd",
            "minLength": 1,
            "maxLength": 200
          },
          "ownerEmail": {
            "type": "string",
            "example": "priya@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery",
            "minLength": 8,
            "maxLength": 200
          }
        },
        "required": [
          "name",
          "ownerEmail",
          "password"
        ]
      },
...
      "TenantRegistrationResponse": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantResponse"
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerUserResponse"
          }
        },
        "required": [
          "tenant",
          "owner"
        ]
      },
...
      "SetupChecklistResponse": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetupChecklistStepResponse"
            }
          }
        },
        "required": [
          "steps"
        ]
      },
...
    }
  }
}